requirements.txt
python-dotenv==1.0.0
requests==2.31.0
pandas==2.1.4
streamlit==1.29.0
plotly==5.18.0
openpyxl==3.1.2
xmltodict==0.13.0
tqdm==4.66.1
